✨ Add graphql-codegen
