+++ codegen.ts
@@ -1,0 +1,13 @@
+import type { CodegenConfig } from "@graphql-codegen/cli";
+
+const config: CodegenConfig = {
+  overwrite: true,
+  schema: "./src/graphql/**/*.graphql",
+  generates: {
+    "src/generated/graphql.ts": {
+      plugins: ["typescript", "typescript-resolvers"],
+    },
+  },
+};
+
+export default config;
