+++ package.json
@@ -5,23 +5,32 @@
   "main": "index.js",
   "type": "module",
   "scripts": {
-    "compile": "tsc",
-    "start": "npm run compile && node ./dist/index.js",
-    "prepare": "husky"
+    "generate": "graphql-codegen --config codegen.yml",
+    "compile": "npm run generate && tsc src/index.ts",
+    "start": "npm run generate && tsx src/index.ts",
+    "dev": "npm run generate && tsx --watch src/index.ts",
+    "prepare": "husky",
+    "codegen": "graphql-codegen --config codegen.ts"
   },
   "author": "Darius Kletter",
   "license": "ISC",
   "dependencies": {
     "apollo-server": "^3.13.0",
+    "bcryptjs": "^2.4.3",
     "graphql": "^16.8.1",
+    "jsonwebtoken": "^9.0.2",
     "mongoose": "^8.2.0"
   },
   "devDependencies": {
-    "@types/node": "^20.11.20",
+    "@graphql-codegen/cli": "5.0.2",
+    "@graphql-codegen/typescript": "4.0.6",
+    "@graphql-codegen/typescript-resolvers": "4.0.6",
+    "@types/node": "^20.11.24",
     "@types/tunnel": "^0.0.7",
     "husky": "^9.0.11",
     "lint-staged": "^15.2.2",
     "prettier": "3.2.5",
+    "tsx": "^4.7.1",
     "tunnel": "^0.0.6",
     "typescript": "^5.3.3"
   },
