+++ src/graphql/resolvers/posts.ts
@@ -1,0 +1,16 @@
+import Post from "../../models/Post.js";
+
+const postsResolver = {
+  Query: {
+    getPosts: async () => {
+      try {
+        const posts = await Post.find();
+        return posts;
+      } catch (err) {
+        throw new Error(err as string);
+      }
+    },
+  },
+};
+
+export default postsResolver;
