+++ src/graphql/schema.graphql
@@ -1,0 +1,28 @@
+type Post {
+  id: ID!
+  body: String!
+  username: String!
+  createdAt: String!
+}
+
+type User {
+  id: ID!
+  email: String!
+  token: String!
+  createdAt: String!
+}
+
+input RegisterInput {
+  username: String!
+  password: String!
+  confirmPassword: String!
+  email: String!
+}
+
+type Query {
+  getPosts: [Post]
+}
+
+type Mutation {
+  register(registerInput: RegisterInput): User!
+}
